src/fast-hamm/requirements.txt
Cython
flask
msgpack
python-rocksdb
numpy
